==> notebooks/1-clean-data.ipynb <==
import pandas as pd
import numpy as np

input_path = "../../dataset/"
output_path = "../data/"

### construct recipe_data

raw_recipes = pd.read_csv(input_path + 'RAW_recipes.csv')
pp_recipes = pd.read_csv(input_path + 'PP_recipes.csv')

raw_recipes.head()

pp_recipes.head()

raw_recipes_data = raw_recipes.loc[:, ['name', 'id', 'nutrition']]
raw_recipes_data.head()

pp_recipes_data = pp_recipes.loc[:, ['id', 'ingredient_ids', 'i']]
pp_recipes_data.head()

# cleaned up racipe dataframe
recipe_data = pd.merge(raw_recipes_data, pp_recipes_data, on = 'id')
recipe_data = recipe_data.rename(columns={'i': 'fid', 'id' : 'full_id'})
recipe_data.set_index('fid', inplace = True)
recipe_data.sort_index(inplace = True)
recipe_data.head()

# [calories (#), total fat (PDV), sugar (PDV) , sodium (PDV) , protein (PDV) , saturated fat (PDV) , and carbohydrates (PDV)]

recipe_data.to_csv(output_path + 'recipe_data.csv')

### construct ingr_data

ingr_map = pd.read_pickle(input_path + 'ingr_map.pkl')

ingr_map.head()

ingr_data = ingr_map.loc[:, ['id', 'replaced', 'count']]
ingr_data.drop_duplicates(inplace=True)
ingr_data.sort_values(by='id', inplace=True)
ingr_data.rename(columns = {'replaced' : 'name', 'id': 'iid'}, inplace = True)
ingr_data.set_index('iid', inplace = True)
ingr_data.sort_index(inplace = True)
ingr_data.head()

ingr_data.to_csv(output_path + 'ingr_data.csv')

### construct rate_data

PP_users = pd.read_csv(input_path + 'PP_users.csv')
rate_data = PP_users.set_index('u').loc[:, ['items', 'ratings']]
rate_data.head()

rate_data['items'] = rate_data['items'].str.replace(" ", "")
rate_data['items'] = rate_data['items'].apply(lambda x: x[1:-1].split(','))

rate_data['ratings'] = rate_data['ratings'].str.replace(" ", "")
rate_data['ratings'] = rate_data['ratings'].apply(lambda x: x[1:-1].split(','))

rate_data.head()

user_list = []
item_list = []
rating_list = []

for index, row in rate_data.iterrows():
    for x in range (0, len(row['items'])):
        item_id = int(row['items'][x])
        user_list.append(index)
        item_list.append(item_id)
        rating_list.append(float(row['ratings'][x]))

rating_data = pd.DataFrame({"user": user_list, "item": item_list, "rate": rating_list})

rating_data.to_csv(output_path + 'rating_data.csv', index = False, header = False)

### construct reduced_rating_data

from surprise import Reader
from surprise import Dataset

# constants
user_size = 1000
item_size = 5000

df = pd.read_csv(output_path + 'rating_data.csv', names = ['userID', 'itemID', 'rating'])
df.head()

reduced = df[(df['userID']<user_size) & (df['itemID']<item_size)]

reduced

reduced.to_csv(output_path + 'reduced_rating_data.csv', index = False, header = False)

reader = Reader(rating_scale=(1, 5))
data = Dataset.load_from_df(reduced[['userID', 'itemID', 'rating']], reader)
