==> notebooks/2-build-mixed-constraints.ipynb <==
import pandas as pd

# constants
# file path
input_path = '../data/'
output_path = '../data/'

# number of dataset of each type of constraint
const_num = 3

# size
user_max = 1000
item_max = 5000

# anchors
anc = [x*(user_max/8) for x in range (0, 9)]

# constraint column names
ul_name = 'u'
i1_name = 'i1'
i2_name = 'i2'
nl_name = 'nl'

def read_cnst(const_num):
    const1 = pd.read_csv(input_path + 'const_1.' + str(const_num) + '.csv', index_col = 'u')
    const2 = pd.read_csv(input_path + 'const_2.' + str(const_num) + '.csv', index_col = 'u')
    const3 = pd.read_csv(input_path + 'const_3.' + str(const_num) + '.csv', index_col = 'u')
    return const1, const2, const3

def make_cnst(const_num, const1, const2, const3):
    ul = []
    i1 = []
    i2 = []
    nl = []

    for u in range(int(anc[0]), int(anc[1])):
        ul.append(u)
        i1.append(const1.iloc[u].i1)
        i2.append(None)
        nl.append(None)

    for u in range(int(anc[1]), int(anc[2])):
        ul.append(u)
        i1.append(None)
        i2.append(const2.iloc[u].i2)
        nl.append(None)

    for u in range(int(anc[2]), int(anc[3])):
        ul.append(u)
        i1.append(None)
        i2.append(None)
        nl.append(const3.iloc[u].nl)

    for u in range(int(anc[3]), int(anc[4])):
        ul.append(u)
        i1.append(const1.iloc[u].i1)
        i2.append(const2.iloc[u].i2)
        nl.append(None)

    for u in range(int(anc[4]), int(anc[5])):
        ul.append(u)
        i1.append(None)
        i2.append(const2.iloc[u].i2)
        nl.append(const3.iloc[u].nl)

    for u in range(int(anc[5]), int(anc[6])):
        ul.append(u)
        i1.append(const1.iloc[u].i1)
        i2.append(None)
        nl.append(const3.iloc[u].nl)

    for u in range(int(anc[6]), int(anc[7])):
        ul.append(u)
        i1.append(const1.iloc[u].i1)
        i2.append(const2.iloc[u].i2)
        nl.append(const3.iloc[u].nl)

    for u in range(int(anc[7]), int(anc[8])):
        ul.append(u)
        i1.append(None)
        i2.append(None)
        nl.append(None)


    data = {
        ul_name: ul,
        i1_name: i1,
        i2_name: i2,
        nl_name: nl
    }

    df = pd.DataFrame(data)
    df.to_csv(output_path + 'const.' + str(const_num) + '.csv', index = False)
    return df

for i in range (1, const_num+1):
    a, b, c = read_cnst(i)
    make_cnst(i, a, b, c)
