==> notebooks/2-build-single-constraint.ipynb <==
import pandas as pd
import copy
import random
import sys
import numpy as np
from numpy import dot
from numpy.linalg import norm

# setting path
sys.path.append('../code/src')

import load_data as ld

# constsant
# file path
dataset_path = "../../dataset/"
input_path = "../data/"
output_path = "../data/"

# size reduction
user_max = 1000
item_max = 5000

# number of dataset of each type of constraint
const_num = 3

# constraint column names
ul_name = 'u'
i1_name = 'i1'
i2_name = 'i2'
nl_name = 'nl'

# parameters for const1/const2
ingr_percent = 1.5
good_rat = 5.0
bad_rat = 0.0

# Load & Edit Dataframes

### Ingredient List

ingr_data = pd.read_csv(output_path + 'ingr_data.csv')
ingr_data.sort_values(by = 'count', ascending = False, inplace = True)
ingr_data.head()

ingr_candidate = ingr_data.head(int(len(ingr_data)*(ingr_percent/100)))
ingr_candidate.head()

ingr_list = ingr_candidate['iid'].tolist()

### Recipe List

recipe_data = ld.load_attr(input_path + 'recipe_data.csv')
recipe_data.head()

### User List

user_data = pd.read_csv(dataset_path + 'PP_users.csv')

user_data['items'] = user_data['items'].str.replace(" ", "")
user_data['items'] = user_data['items'].apply(lambda x: x[1:-1].split(','))
user_data['ratings'] = user_data['ratings'].str.replace(" ", "")
user_data['ratings'] = user_data['ratings'].apply(lambda x: x[1:-1].split(','))

user_data = user_data[user_data['u'] < user_max]
user_data.set_index('u', inplace = True)
user_data = user_data.loc[:, ['items', 'ratings']]

user_data

# Construct Constraint 1 and 2

### Helper Function

def random_select(x): # pick from 0 to x
    result = []
    while len(result)<const_num:
        p = int(random.uniform(0, x))
        if p not in result:
            result.append(p)
    return result

### Consistency of Constraint
ingredient related constraints need to be consistent  
ex) included ingredient shouldn't have bad rating  
ex) excluded ingredient shouldn't have good rating

ingr_column = [copy.deepcopy(ingr_list) for x in range(len(user_data))]
len(ingr_column)

user_data = user_data.assign(include = ingr_column)
user_data = user_data.assign(exclude = copy.deepcopy(ingr_column))
user_data.head()

for index, row in user_data.iterrows():
    for x in range (0, len(row['items'])):
        item_id = int(row['items'][x])
        if(item_id >= item_max): # ignore if item id is larger or equal than 5000
            continue
            
        if float(row['ratings'][x]) == good_rat : # remove ingredients included in the item frome exclude list
            ingrs = recipe_data.loc[item_id].ingredient_ids
            for ingr in ingrs:
                if int(ingr) in row['exclude']:
                    row['exclude'].remove(int(ingr))
        if float(row['ratings'][x]) == bad_rat :
            ingrs = recipe_data.loc[item_id].ingredient_ids
            for ingr in ingrs:
                if int(ingr) in row['include']:
                    row['include'].remove(int(ingr))

user_data.head()

for index, row in user_data.iterrows():
    # select random ingr to include
    if len(row['include']) > const_num:
        pick = random_select(len(row['include'])-1)
        result = []
        for p in pick:
            result.append(row['include'][p])
        row['include'] = result

    # select random ingr to exclude
    if len(row['exclude']) > const_num:
        pick = random_select(len(row['exclude'])-1)
        result = []
        for p in range (0, len(pick)):
            result.append(row['exclude'][p])
        row['exclude'] = result            

user_data.head()

ingr_const = user_data.loc[:, ['include', 'exclude']]
ingr_const.head()

## Construct const_1.x

# build 3 datasets
for i in range (0, const_num):
    ul = []
    i1 = []
    i2 = []
    nl = []
    
    for u in range (0, user_max):
        ul.append(u)
        i1.append(ingr_const.iloc[u].include[i])
        i2.append(None)
        nl.append(None)
    
    data = {
        ul_name: ul,
        i1_name: i1,
        i2_name: i2,
        nl_name: nl
    }
    
    df = pd.DataFrame(data)
    df.to_csv(output_path + 'const_1.' + str(i+1) + '.csv', index = False)

df.head()

## Construct Const_2.x

# build 3 datasets
for i in range (0, const_num):
    ul = []
    i1 = []
    i2 = []
    nl = []
    
    for u in range (0, user_max):
        ul.append(u)
        i2.append(ingr_const.iloc[u].exclude[i])
        i1.append(None)
        nl.append(None)
    
    data = {
        ul_name: ul,
        i1_name: i1,
        i2_name: i2,
        nl_name: nl
    }
    
    df = pd.DataFrame(data)
    df.to_csv(output_path + 'const_2.' + str(i+1) + '.csv', index = False)

df.head()

# Construct Nutrient Related Constraints

### Helper Function

def cos_sim(A, B):
  return dot(A, B)/(norm(A)*norm(B))

c3_dict = {}
for user in range (0, user_max):
    nutrs= []
    for item in user_data.iloc[user][0]:
        idx = user_data.iloc[user][0].index(item)
        if int(item) > item_max: # ignore if item idx > item_max
            continue
        if float(user_data.iloc[user][1][idx]) < 4: # ignore if item rate < 4
            continue
        nutrs.append(recipe_data.iloc[int(item)].nutrition)
    nutrs.sort()
    
    # pick 3 nutrients
    c3_dict[user] = []
    if len(nutrs) == 0: # no information
        for i in range (0, const_num):
            p = int(random.uniform(0, item_max))
            c3_dict[user].append(recipe_data.iloc[p].nutrition)
    else:
        c3_dict[user].append(nutrs[0])
        c3_dict[user].append(nutrs[int(len(nutrs)/2)])
        c3_dict[user].append(nutrs[-1])

### Test

alp = 0.5
user = 3
for i in range (2, const_num):
    RMSE = 0
    for item in user_data.iloc[user][0]:
        idx = user_data.iloc[user][0].index(item)
        if int(item) > item_max:
            continue
        rate = float(user_data.iloc[user][1][idx])
        if rate < 4:
            continue
        nutr = recipe_data.iloc[int(item)].nutrition
        
        new_rate = cos_sim(nutr, c3_dict[user][i]) * alp * 5
        new_rate = new_rate + rate * (1-alp)
        RMSE = RMSE + (new_rate - rate) * (new_rate - rate)
        print('rate: ', rate)
        print('similarity: ', cos_sim(nutr, c3_dict[user][i]) * alp * 5)
        print('new rate: ', new_rate)
    RMSE = RMSE / len(user_data.iloc[user][0])
    print(RMSE)
    break

## Construct Const_3.x

# build 3 datasets
for i in range (0, const_num):
    ul = []
    i1 = []
    i2 = []
    nl = []
    
    for u in range (0, user_max):
        ul.append(u)
        i1.append(None)
        i2.append(None)
        nl.append(c3_dict[u][i])
    
    data = {
        ul_name: ul,
        i1_name: i1,
        i2_name: i2,
        nl_name: nl
    }
    
    df = pd.DataFrame(data)
    df.to_csv(output_path + 'const_3.' + str(i+1) + '.csv', index = False)
    df.head()
