==> notebooks/3-sample.ipynb <==
import pandas as pd

# constants
# file path
result_path = '../result/'

# Post-Rec Single

### Constraint 1

pd.read_csv(result_path + 'PostRec_1.1.csv', index_col=0)

### Constraint 2

pd.read_csv(result_path + 'PostRec_2.1.csv', index_col=0)

### Constraint 3

pd.read_csv(result_path + 'PostRec_3.1.csv', index_col=0)

# Post-Rec Mixed

pd.read_csv(result_path + 'PostRec.1.csv', index_col=0)

# Inter-Rec Single

### Constraint 1

pd.read_csv(result_path + 'InterRec_1.1.csv', index_col=0).sort_index()

### Constraint 2

pd.read_csv(result_path + 'InterRec_2.1.csv', index_col=0).sort_index()

### Constraint 3

pd.read_csv(result_path + 'InterRec_3.1.csv', index_col=0).sort_index()

# Inter-Rec Mixed

pd.read_csv(result_path + 'InterRec.1.csv', index_col=0).sort_index()
