==> notebooks/5-CnstSVD-analysis.ipynb <==
# import modules
import copy
import time
from sys import path
from collections import defaultdict

import pandas as pd
from surprise import SVD, Reader, Dataset, accuracy
from surprise.model_selection import train_test_split

path.append('../code/src')

import svd_constraint
from inter_rec import InterRec
from evaluate import Evaluation as ev

# constants
# file path
input_path = "../data/"
output_path = '../result/'
rate_file = 'reduced_rating_data.csv'
attr_file = 'recipe_data.csv'

# user/item max
user_max = 1000
food_max = 5000

# const count
const_count = 3

# Dictionary keys
rs1 = 'CnstSVD'
rs2 = 'CnstSVD-all'
rs3 = 'CnstSVD-hard'
rs4 = 'CnstSVD-harder'
val_r1 = 'RMSE'
val_r2 = 'nDCG'
val_t1 = 'train time(s)'
val_t2 = 'exec time(s)'

val_dict = {
    val_r1: 0,
    val_r2: 0,
    val_t1: 0,
    val_t2: 0
}

keys = [rs1]
keys = keys + [rs2]
keys = keys + [rs3]
keys = keys + [rs4]

result = {key: copy.deepcopy(val_dict) for key in keys}

# read previous result
result_df = pd.read_csv(output_path + 'evaluation_SVD.csv', index_col = 0).transpose()
result.update(result_df.to_dict())
result

# Evaluate CnstSVD

def run_inter(idx, algo, k):    
    # InterRec applies constraint after the rating of each item is predicted
    rec = InterRec(input_path + rate_file, input_path + attr_file, 
                           input_path + 'const' + '.' + str(idx) + '.csv', 
                            algo, need_test = True)
    rec.set_result_n(k)
    rec.get_data()  # get rating, attribute, recipe data
    
    # train with data
    start = time.time()
    rec.train()  
    t1 = time.time() - start
    
    # predict rating for test-set
    predict_test = rec.test_rmse()
    r1 = ev.calculate_rmse(predict_test)
    
    # get top-n for anti-test-set
    start = time.time()
    rec.test()
    top_n_df = rec.get_top_n()
    t2 = time.time() - start

    # calculate nDCG
    r2 = ev.calculate_ndcg(rec.get_rel(), top_n_df, k)
        
    return r1, r2, t1, t2

r1_sum = 0
r2_sum = 0
t1_sum = 0
t2_sum = 0
for j in range (1, const_count + 1):
    r1, r2, t1, t2 = run_inter(j, svd_constraint.CnstSVD(), 50)
    r1_sum = r1_sum + r1
    r2_sum = r2_sum + r2
    t1_sum = t1_sum + t1
    t2_sum = t2_sum + t2
    print('Const'+'.'+str(j)+" done")

result[rs1][val_r1] = r1_sum/const_count
result[rs1][val_r2] = r2_sum/const_count
result[rs1][val_t1] = t1_sum/const_count
result[rs1][val_t2] = t2_sum/const_count

# Evaluate CnstSVD_all

r1_sum = 0
r2_sum = 0
t1_sum = 0
t2_sum = 0
for j in range (1, const_count + 1):
    r1, r2, t1, t2 = run_inter(j, svd_constraint.CnstSVD_all(), 50)
    r1_sum = r1_sum + r1
    r2_sum = r2_sum + r2
    t1_sum = t1_sum + t1
    t2_sum = t2_sum + t2
    print('Const'+'.'+str(j)+" done")

result[rs2][val_r1] = r1_sum/const_count
result[rs2][val_r2] = r2_sum/const_count
result[rs2][val_t1] = t1_sum/const_count
result[rs2][val_t2] = t2_sum/const_count

# Evaluate CnstSVD_hard

r1_sum = 0
r2_sum = 0
t1_sum = 0
t2_sum = 0
for j in range (1, const_count + 1):
    r1, r2, t1, t2 = run_inter(j, svd_constraint.CnstSVD_hard(), 50)
    r1_sum = r1_sum + r1
    r2_sum = r2_sum + r2
    t1_sum = t1_sum + t1
    t2_sum = t2_sum + t2
    print('Const'+'.'+str(j)+" done")

result[rs3][val_r1] = r1_sum/const_count
result[rs3][val_r2] = r2_sum/const_count
result[rs3][val_t1] = t1_sum/const_count
result[rs3][val_t2] = t2_sum/const_count

# Evaluate CnstSVD_harder

r1_sum = 0
r2_sum = 0
t1_sum = 0
t2_sum = 0
for j in range (1, const_count + 1):
    r1, r2, t1, t2 = run_inter(j, svd_constraint.CnstSVD_harder(), 50)
    r1_sum = r1_sum + r1
    r2_sum = r2_sum + r2
    t1_sum = t1_sum + t1
    t2_sum = t2_sum + t2
    print('Const'+'.'+str(j)+" done")

result[rs4][val_r1] = r1_sum/const_count
result[rs4][val_r2] = r2_sum/const_count
result[rs4][val_t1] = t1_sum/const_count
result[rs4][val_t2] = t2_sum/const_count

# Result

result_df = pd.DataFrame.from_dict(result, orient='index', columns=['RMSE', 'nDCG', 'train time(s)', 'exec time(s)'])
result_df.to_csv(output_path + 'evaluation_SVD.csv')
result_df
